Add dummy ORDER BY as needed for MSSQL limits
SQL Server only supports offset and limit functionality as a part of
their ORDER BY syntax. This means if a query doesn't have an order
clause we will generate invalid syntax when setting a limit. To avoid
this when a limit or offset is requested and no order is given we add
a "do nothing" ORDER BY clause. This should have no effect on the
query performance or the order of the results.