fix: STI types on children in joins (#3160)
* Use most specific matching relation type

This allows using a single table for multiple entities without using a
type column. Some setups infer the type from the context of how/where
the row is loaded.

* Use entity's relation metadata for joins

If an STI child overrides a parent property and uses a different type
doing a query on the parent type will skip the relation. This happens
because the query's join is built using the parent relation but when
mapping to an entity the child's relation is fetched so they don't
match. Instead we now use the relation's propertyPath to allow using
relations that reference the same property. We also copy the child's
relation type into the join to ensure we get the right type in the child
as well.